requirements.txt
# Dependencias de la Plataforma de Evaluación de Asertividad

# Framework principal
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0

# Utilidades y seguridad
Werkzeug==3.0.1
requests==2.32.4

# Servidor de producción
gunicorn==21.2.0

# Desarrollo y testing (opcional)
# pytest==7.4.3
# pytest-flask==1.3.0
